+++ src/dom_handler.js
@@ -1,7 +1,23 @@
-import { setListener } from './events_handler';
-import { projects } from './variables';
+// import { setListener } from './events_handler';
+import projects from './variables';
 
-export function renderProjects() {
+function describeProject(e) {
+  const index = e.target.id.substring(e.target.id.length - 1, e.target.id.length);
+  const project = projects[index];
+  document.getElementById('title-project').innerHTML = project.name;
+  document.getElementById('desc-project').innerHTML = project.description;
+  // console.log(project);
+}
+
+function setListener() {
+  projects.forEach((elem, index) => {
+    document
+      .getElementById(`project-n-${index}`)
+      .addEventListener('click', describeProject);
+  });
+}
+
+export default function renderProjects() {
   const listNode = document.getElementById('projects-list');
   listNode.innerHTML = '';
   projects.forEach((elem, index) => {
