+++ src/index.js
@@ -1,7 +1,7 @@
 import Project from './classes/project';
-import { addListenerToProjects } from './events_handler';
-import { projects } from './variables';
-import { renderProjects } from './dom_handler';
+import addListenerToProjects from './events_handler';
+import projects from './variables';
+import renderProjects from './dom_handler';
 
 const project1 = new Project('Work', 'Work description');
 const project2 = new Project('Office', 'Office description');
